server/requirements.txt
django-cors-headers
django-rest
django
python-dotenv